Cleanups and fixes
Drop the -bin package since a plugin is not a user executable bin, also,
don't split into a separate package since a plugin will most certainly
not work without the right lib it was built against, and using a lib
without its plugin is a bit useless.

Other misc. changes include not depending on qt5-default but instead
exporting the correct var for qtchooser and reintroducing the data
package.